requirements.txt
streamlit>=1.31.0
openai>=1.12.0
wikipedia-api>=0.6.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
sqlite3>=3.35.0
coverage>=7.4.0
pytest>=8.0.0
pytest-cov>=4.1.0
tqdm==4.66.2
tabulate==0.9.0
langchain==0.1.12
langchain-openai==0.0.8
langchain-community==0.0.27 